requirements.txt
Flask==3.0.0
py3_validate_email==1.0.5.post1
